+++ ml_model_manager.py
@@ -63,8 +63,8 @@ class MLModelManager:
         model_names = [model.get("id", "") for model in available_models]
         model_states = [model.get("state", "") for model in available_models]
 
-        if model_name not in model_names or model_states[model_names.index(model_name)] != "loaded":
-            raise ValueError(f"Model '{model_name}' not available or loaded")
+        # if model_name not in model_names or model_states[model_names.index(model_name)] != "loaded":
+        #     raise ValueError(f"Model '{model_name}' not available or loaded")
 
         self.model = model_name
         self.system_prompt = system_prompt
